requirements.txt
#run "docker-compose build" everytime dependency is added to rebuild the docker image

Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
uwsgi>=2.0.20<2.1
phonenumbers
djangorestframework_simplejwt==4.7.2
requests
PyJWT
python-decouple
python-dotenv
cryptography